# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OnboardingAnswers": {
        "description": "Raw answers collected from the onboarding flow.",
        "properties": {
          "checkFrequency": {
            "default": "",
            "title": "Checkfrequency",
            "type": "string"
          },
          "conservativeDefault": {
            "default": "",
            "title": "Conservativedefault",
            "type": "string"
          },
          "country": {
            "default": "",
            "title": "Country",
            "type": "string"
          },
          "cryptoComfort": {
            "default": "",
            "title": "Cryptocomfort",
            "type": "string"
          },
          "currency": {
            "default": "",
            "title": "Currency",
            "type": "string"
          },
          "debtStress": {
            "default": "",
            "title": "Debtstress",
            "type": "string"
          },
          "debtTypes": {
            "items": {
              "type": "string"
            },
            "title": "Debttypes",
            "type": "array"
          },
          "dependents": {
            "default": "",
            "title": "Dependents",
            "type": "string"
          },
          "emergencyFundMonths": {
            "default": "",
            "title": "Emergencyfundmonths",
            "type": "string"
          },
          "existingInvestments": {
            "items": {
              "type": "string"
            },
            "title": "Existinginvestments",
            "type": "array"
          },
          "incomePredictability": {
            "default": "",
            "title": "Incomepredictability",
            "type": "string"
          },
          "investedAmountBand": {
            "default": "",
            "title": "Investedamountband",
            "type": "string"
          },
          "investingStyle": {
            "default": "",
            "title": "Investingstyle",
            "type": "string"
          },
          "knowledgeLevel": {
            "default": "",
            "title": "Knowledgelevel",
            "type": "string"
          },
          "lifestyleLossCapacity": {
            "default": "",
            "title": "Lifestylelosscapacity",
            "type": "string"
          },
          "lossReaction": {
            "default": "",
            "title": "Lossreaction",
            "type": "string"
          },
          "majorExpenses": {
            "default": "",
            "title": "Majorexpenses",
            "type": "string"
          },
          "monthlyDisposableBand": {
            "default": "",
            "title": "Monthlydisposableband",
            "type": "string"
          },
          "monthlyIncomeBand": {
            "default": "",
            "title": "Monthlyincomeband",
            "type": "string"
          },
          "preferredCompanies": {
            "items": {
              "type": "string"
            },
            "title": "Preferredcompanies",
            "type": "array"
          },
          "preferredMarkets": {
            "items": {
              "type": "string"
            },
            "title": "Preferredmarkets",
            "type": "array"
          },
          "primaryGoal": {
            "default": "",
            "title": "Primarygoal",
            "type": "string"
          },
          "priorInvesting": {
            "default": "",
            "title": "Priorinvesting",
            "type": "string"
          },
          "priority": {
            "default": "",
            "title": "Priority",
            "type": "string"
          },
          "restrictedSectors": {
            "items": {
              "type": "string"
            },
            "title": "Restrictedsectors",
            "type": "array"
          },
          "timeHorizon": {
            "default": "",
            "title": "Timehorizon",
            "type": "string"
          },
          "volatileStockResponse": {
            "default": "",
            "title": "Volatilestockresponse",
            "type": "string"
          },
          "yearDownReaction": {
            "default": "",
            "title": "Yeardownreaction",
            "type": "string"
          }
        },
        "title": "OnboardingAnswers",
        "type": "object"
      },
      "OnboardingSubmitRequest": {
        "description": "Persisted onboarding submission.",
        "properties": {
          "answers": {
            "$ref": "#/components/schemas/OnboardingAnswers"
          }
        },
        "required": [
          "answers"
        ],
        "title": "OnboardingSubmitRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Scope Stock Research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/onboarding/profile": {
      "get": {
        "description": "Return the current user's onboarding profile.",
        "operationId": "get_profile_api_v1_onboarding_profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "scope_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Profile Api V1 Onboarding Profile Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "onboarding"
        ]
      },
      "post": {
        "description": "Save answers and derived financial/risk profile summaries.",
        "operationId": "save_profile_api_v1_onboarding_profile_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "scope_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Profile Api V1 Onboarding Profile Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Profile",
        "tags": [
          "onboarding"
        ]
      }
    },
...
  }
}